requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pdfminer.six
spacy
numpy
groq
python-dotenv